Logged in user delete and update contact

+++ controllers/contactController.js
@@ -50,6 +50,12 @@ const updateContact = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error("Contact not found");
     }
+
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("USer don`t have permission to update other user contacts");
+    }
+
     const updatedContact = await Contact.findByIdAndUpdate(
         req.params.id,
         req.body,
@@ -67,8 +73,12 @@ const deleteContact = asyncHandler(async (req, res) => {
         res.status(404);
         throw new Error("Contact not found");
     }
+    if (contact.user_id.toString() !== req.user.id) {
+        res.status(403);
+        throw new Error("USer don`t have permission to update other user contacts");
+    }
     await Contact.deleteOne({ _id: req.params.id });
-    res.status(200).json(contact);
+  res.status(200).json(contact);
 });
 
 module.exports = { getContacts, createContact, getContact, updateContact, deleteContact }
